+++ junit-docker/src/main/java/no/mnemonic/commons/junit/docker/MariaDBDockerResource.java
@@ -45,12 +45,13 @@ public class MariaDBDockerResource extends DockerResource {
                                 String exposedPortsRange,
                                 int reachabilityTimeout,
                                 boolean skipReachabilityCheck,
+                                boolean skipPullDockerImage,
                                 Supplier<DockerClient> dockerClientResolver,
                                 Map<String, String> environmentVariables,
                                 String setupScript,
                                 String truncateScript) {
     super(imageName, applicationPorts, exposedPortsRange, reachabilityTimeout, skipReachabilityCheck,
-            dockerClientResolver, environmentVariables);
+            skipPullDockerImage, dockerClientResolver, environmentVariables);
     // Both parameters are optional.
     this.setupScript = !StringUtils.isBlank(setupScript) ? checkFileExists(setupScript) : null;
     this.truncateScript = !StringUtils.isBlank(truncateScript) ? checkFileExists(truncateScript) : null;
@@ -181,7 +182,7 @@ public class MariaDBDockerResource extends DockerResource {
       addEnvironmentVariable("MYSQL_ROOT_PASSWORD", "root");
       addEnvironmentVariable("MYSQL_ROOT_HOST", "%");
       return new MariaDBDockerResource(imageName, applicationPorts, exposedPortsRange, reachabilityTimeout,
-              skipReachabilityCheck, dockerClientResolver, environmentVariables, setupScript, truncateScript);
+              skipReachabilityCheck, skipPullDockerImage, dockerClientResolver, environmentVariables, setupScript, truncateScript);
     }
 
     /**